Refactor method
